TECH-349: Menu type organization
